[update] integration / テスト前に特にファイル削除する必要がなければNoneを指定できるように

+++ test/webserv/integration/test_http_post.py
@@ -13,6 +13,9 @@ UPLOAD_FILE_PATH = UPLOAD_DIR + "test_upload_file"
 
 
 def delete_file(file_path):
+    if file_path is None:
+        return
+
     try:
         if os.path.exists(file_path):
             os.remove(file_path)
@@ -22,6 +25,7 @@ def delete_file(file_path):
         assert False
 
 
+# テスト前に削除しておきたいupload_file_pathが特になければNoneを指定する
 @pytest.mark.parametrize(
     "request_file, expected_response, upload_file_path",
     [
@@ -36,6 +40,10 @@ def delete_file(file_path):
     ],
 )
 def test_post_responses(request_file, expected_response, upload_file_path):
+    # testで作成したいファイルがあるならtest前に念のため削除
     delete_file(upload_file_path)
+
     send_request_and_assert_response(request_file, expected_response)
+
+    # testで作成したファイルがあれば次のtestのために削除
     delete_file(upload_file_path)
